fixed warnings

+++ memory.c
@@ -114,8 +114,8 @@ int var_allocate(char name , int size)
                         }
 
                         //update variables
-                        var_table[name].start = start_free_node;
-                        var_table[name].size = size;
+                        var_table[(int)name].start = start_free_node;
+                        var_table[(int)name].size = size;
 
                         return 1;
                 }
@@ -132,7 +132,7 @@ int var_free(char name)
         FreeBlock * new_free_node;
 
         //if the variable already doesnt exist , cannot free it
-        if(var_table[name].start == -1) return 0;
+        if(var_table[(int)name].start == -1) return 0;
 
         new_free_node = malloc (sizeof(struct FreeBlock));
 
@@ -143,11 +143,11 @@ int var_free(char name)
         previous = NULL;
 
         //set new free node in the list based on the freed variable
-        new_free_node -> start  = var_table[name].start;
-        new_free_node -> size = var_table[name].size;
+        new_free_node -> start  = var_table[(int)name].start;
+        new_free_node -> size = var_table[(int)name].size;
 
         //variable doesnt exist
-        var_table[name].start = -1;
+        var_table[(int)name].start = -1;
 
         // if freelist is empty already
         if (freelist == NULL)
